% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestBus.R
\name{best_length_diff}
\alias{best_length_diff}
\title{Time difference between 3 different travels}
\usage{
best_length_diff(prev_travel, curr_travel, next_travel, best_travel)
}
\arguments{
\item{prev_travel}{length of the previous travel in minutes}

\item{curr_travel}{length of the current travel in minutes}

\item{next_travel}{length of the next travel in minutes}

\item{best_travel}{best travel -1 or 0 or 1}
}
\value{
This function returns the difference between the current travel and the minimum of the 3 different travels.
}
\description{
This function calculates the difference between the current travel and the best travel.
It considers that the function best_travel has already been used to calculate the best_travel.
}
\examples{
best_length_diff(20,30,45,-1)
best_length_diff(35,30,25,1)
best_length_diff(35,30,31,0)
}

