% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestBus.R
\name{best_length_diff_deprecated}
\alias{best_length_diff_deprecated}
\title{Time difference between 3 consecutive travels - deprecated}
\usage{
best_length_diff_deprecated(prev_travel, curr_travel, later_travel)
}
\arguments{
\item{prev_travel}{length of the previous travel in minutes}

\item{curr_travel}{length of the current travel in minutes}

\item{later_travel}{length of the next travel in minutes}
}
\value{
This function returns the difference between the current travel and the minimum of the specifieds travels.
}
\description{
This function calculates the difference between the current travel and the best travel.
}
\examples{
best_length_diff_deprecated(35,100,40)
best_length_diff_deprecated(NA,100,40)
best_length_diff_deprecated(35,100,NA)
}

