% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestBus.R
\name{best_travel}
\alias{best_travel}
\title{The best travel function}
\usage{
best_travel(prev_travel, curr_travel, next_travel, diff_prev_schedule,
  diff_next_schedule, duration_limit, time_limit)
}
\arguments{
\item{prev_travel}{length of the previous travel in minutes}

\item{curr_travel}{length of the main travel in minutes}

\item{next_travel}{length of the next travel in minutes}

\item{diff_prev_schedule}{difference in minutes between the current travel and the previous}

\item{diff_next_schedule}{difference in minutes between the current travel and the next}

\item{duration_limit}{time limit applied on the difference between the best travel duration and the current travel duration}

\item{time_limit}{headway maximum time in order that a travel can be considered the best}
}
\value{
-1 or 0 or 1
}
\description{
This function calculates the index of the minimum travel as long as
the difference between the best travel and the current don't extrapolate
the time limit and the interval between the current travel schedule and
the best travel schedule
}
\examples{
best_travel(50,32,20,25,35,5,45)
best_travel(50,32,20,25,35,5,30)
best_travel(26,32,20,25,35,5,30)
}

