% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestBus.R
\name{group_minutes}
\alias{group_minutes}
\title{Group minutes}
\usage{
group_minutes(schedule, delimiter)
}
\arguments{
\item{schedule}{schedule to be decreased in the format "hh:mm:ss"}

\item{delimiter}{amount of minutes to delimit the schedule}
}
\value{
the schedule descreased
}
\description{
This function is similar to the floor function, but instead of an integer, it is passed a schedule.
The value will be decreased acording to the passed delimiter
}
\examples{
group_minutes("02:17:05", 15)
group_minutes("02:17:05", 20)
}

